Correct paths
